[udev rules] Update rules to use ENV{ID_SOFTWARE_RADIO}="1" for uaccess
This allows the rule to be processed by the uaccess rule and makes the
device accessible by the currently logged in user without needing to be
part of any specific group. For the uaccess rule to work the rules also
needed to be renamed from 88-nuand-xxx to 60-nuand-xxx.

For scripted use without logging the user still needs to be added to
a specific group (plugdev by default).

This is aligned with how it is now done on other SDR devices.

Signed-off-by: Martin Herren (HB9FXX) <cq@hb9fxx.ch>